@@ -1,45 +1,40 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { FormsModule, ReactiveFormsModule} from '@angular/forms';
-import { HTTP_INTERCEPTORS, HttpClientModule } from "@angular/common/http";
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { NgSelectizeModule } from 'ng-selectize';
-import { ToastrModule } from "ngx-toastr";
+import { ToastrModule } from 'ngx-toastr';
 
-import { AuthInterceptor } from "./interceptors/auth.interceptor";
+import { AuthInterceptor } from './interceptors/auth.interceptor';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { PhoneMaskDirective } from './directives/phone-mask/phone-mask.directive';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    PhoneMaskDirective
-  ],
+  declarations: [AppComponent, PhoneMaskDirective],
   imports: [
     AppRoutingModule,
     BrowserModule,
     NgbModule,
     HttpClientModule,
     BrowserAnimationsModule,
-    NgSelectizeModule,
     FormsModule,
     ReactiveFormsModule,
     ToastrModule.forRoot({
       timeOut: 5000,
-      positionClass: "toast-bottom-right",
+      positionClass: 'toast-bottom-right',
       preventDuplicates: true,
-    })
+    }),
   ],
-  exports: [ PhoneMaskDirective ],
+  exports: [PhoneMaskDirective],
   providers: [
     {
       provide: HTTP_INTERCEPTORS,
       useClass: AuthInterceptor,
       multi: true,
-    }
+    },
   ],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}
